Contour plot

A plotting function for visualizing likelihood surface for a pair of parameters using OUwie.contour data.

# S3 method for OUwie.contour
plot(x, mle.point=NULL, levels=c(0:20*0.1), xlab=NULL, 
ylab=NULL, xlim=NULL, ylim=NULL, col=grey.colors(21, start=0, end=1), ...)

Arguments

x

a OUwie.contour object.

mle.point

specifies the color for the maximum likelihood set of parameters. By default the point is not plotted (i.e., set to NULL.

levels

the levels at which to draw contour lines, measured as lnL units away from the best values.

xlab

allows users to specify the x-axis label.

ylab

allows users to specify the y-axis label.

xlim

allows users to specify the lower and upper limits of the x-axis.

ylim

allows users to specify the lower and upper limits of the y-axis.

col

indicates the color gradient.

...

additional parameters to control the plot.

References

Beaulieu J.M., Jhwueng D.C., Boettiger C., and O'Meara B.C. 2012. Modeling stabilizing selection: Expanding the Ornstein-Uhlenbeck model of adaptive evolution. Evolution 66:2369-2383.

Author

Jeremy M. Beaulieu